requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-apscheduler==1.13.1
pytrends==4.9.2
feedparser==6.0.11
googlenewsdecoder==0.1.2
requests==2.32.3
beautifulsoup4==4.12.3
newspaper3k==0.2.8
lxml==5.3.0
python-dotenv==1.0.1
gunicorn==23.0.0
